DefaultSettingsPropertyFactoryDelegate.swift
