+++ wire-ios/Wire-iOS/Sources/UserInterface/SelfProfile/SelfProfileViewController.swift
@@ -229,76 +229,3 @@ extension SelfProfileViewController: AccountSelectorViewDelegate {
         }
     }
 }
-
-// TODO: move to separate file
-
-final class DefaultSettingsPropertyFactoryDelegate: SettingsPropertyFactoryDelegate, PasscodeSetupViewControllerDelegate {
-
-    var userSession: any UserSession
-    var settingsTableViewController: () -> SettingsTableViewController?
-    var mainCoordinator: AnyMainCoordinator<MainCoordinatorDependencies>
-
-    init(
-        userSession: any UserSession,
-        settingsTableViewController: @escaping () -> SettingsTableViewController?,
-        mainCoordinator: AnyMainCoordinator<MainCoordinatorDependencies>
-    ) {
-        self.userSession = userSession
-        self.settingsTableViewController = settingsTableViewController
-        self.mainCoordinator = mainCoordinator
-    }
-
-    /// Create or delete custom passcode when appLock option did change
-    /// If custom passcode is not enabled, no action is taken
-    ///
-    /// - Parameters:
-    ///   - settingsPropertyFactory: caller of this delegate method
-    ///   - newValue: new value of app lock option
-    ///   - callback: callback for PasscodeSetupViewController
-    func appLockOptionDidChange(
-        _ settingsPropertyFactory: SettingsPropertyFactory,
-        newValue: Bool,
-        callback: @escaping ResultHandler
-    ) {
-        // There is an additional check for the simulator because there's no way to disable the device passcode on the simulator. We need it for testing.
-        guard AuthenticationType.current == .unavailable || (UIDevice.isSimulator && AuthenticationType.current == .passcode) else {
-            callback(newValue)
-            return
-        }
-
-        guard newValue else {
-            try? userSession.deleteAppLockPasscode()
-            callback(newValue)
-            return
-        }
-
-        let passcodeSetupViewController = PasscodeSetupViewController(context: .createPasscode, callback: callback)
-        passcodeSetupViewController.passcodeSetupViewControllerDelegate = self
-
-        let keyboardAvoidingViewController = KeyboardAvoidingViewController(viewController: passcodeSetupViewController)
-
-        let wrappedViewController = keyboardAvoidingViewController.wrapInNavigationController(navigationBarClass: TransparentNavigationBar.self)
-
-        let closeItem = passcodeSetupViewController.closeItem
-
-        keyboardAvoidingViewController.navigationItem.leftBarButtonItem = closeItem
-
-        wrappedViewController.presentationController?.delegate = passcodeSetupViewController
-
-        Task {
-            // TODO: test on iPad and iPhone
-            await mainCoordinator.presentViewController(wrappedViewController)
-        }
-    }
-
-    // MARK: - PasscodeSetupViewControllerDelegate
-
-    func passcodeSetupControllerDidFinish() {
-        // no-op
-    }
-
-    func passcodeSetupControllerWasDismissed() {
-        // refresh options applock switch
-        settingsTableViewController()?.refreshData()
-    }
-}
